+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Warn if default author name is in use; OK for testing, not allowed in production
 - Automatically discover sequenza path, oncotree code, tumour id in `cnv` plugin
 - Further simplification of INI parameters for `cnv` and `snv_indel` plugins
+- Check purity is consistent between `input_params_helper` and `cnv` plugin
 
 ## v1.0.0-dev0.0.12: 2023-10-19